refactor(ingestion): registry owns ETA and reporting; dissolve ProgressReporter
