+++ ingestion/tests/unit/lineage/test_snowflake_access_history_progress.py
@@ -14,7 +14,6 @@ from unittest.mock import MagicMock
 
 from metadata.ingestion.api.models import Either
 from metadata.ingestion.source.database.snowflake.lineage import SnowflakeLineageSource
-from metadata.workflow.progress_render import ProgressReporter
 
 
 class TestAccessHistoryProgress:
@@ -29,4 +28,4 @@ class TestAccessHistoryProgress:
 
         assert len(results) == 4
         assert source.progress.global_counters() == [("LineageRecords", 4, None)]
-        assert ProgressReporter(source.progress).eta_seconds() is None
+        assert source.progress.eta_seconds() is None
